backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["rmsdr <romainsadar.pro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2"
djangorestframework = "^3.16.0"
django-cors-headers = "^4.7.0"
psycopg2-binary = "^2.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
